Gym3RenderWrapper should inherit from gym.Wrapper instead
