+++ alf/environments/gym_wrappers.py
@@ -689,3 +689,55 @@ class NonEpisodicEnv(gym.Wrapper):
     def step(self, action):
         ob, reward, done, info = self.env.step(action)
         return ob, reward, False, info
+
+
+class Gym3RenderWrapper(gym.Wrapper):
+    """Wraps gym-based Gym3 environments to enable rendering
+
+    The set of Gym3 environemtns provide a gym wrapper which is mostly
+    compatible with existing gym interfaces except for rendering.
+
+    The render is enabled upon constructing the environments instead of by
+    calling ``render()``. In fact for Gym3 environments ``render()`` does
+    nothing.
+
+    This wrapper works that around by re-constructing the gym-based Gym3
+    environment upon the first call to ``render()``, with specified mode. This
+    makes it transparent for alf's play.
+
+    Note that in our use case ``render()`` is almost always called at
+    the very beginning of a play. This makes reconstructing the
+    environment a plausible approach that does not interrupt ongoing
+    episodes.
+
+    """
+
+    def __init__(self, gym_env):
+        """
+        Args:
+            gym_env (gym3.interop.ToGymEnv): An instance of OpenAI gym environment.
+
+        """
+        import gym3
+        assert isinstance(gym_env, gym3.interop.ToGymEnv), \
+            f'Gym3RenderWrapper requires gym3 wrapped environment, not {type(gym_env)}'
+
+        super().__init__(gym_env)
+
+        # Once enabled (set to True), it will be True forever.
+        self._render_enabled = False
+
+    def render(self, mode: str):
+        """Enables rendering by reconstructing the environment
+        """
+        if not self._render_enabled:
+            self.__init__(
+                self.env.spec.make(render=(mode == 'human'), render_mode=mode))
+            self._render_enabled = True
+
+        # For rgb_array rendering, return the image frame
+        wrapped_env = self.env.env
+        if len(wrapped_env.get_info()) > 0:
+            if 'rgb' in wrapped_env.get_info()[0]:
+                return wrapped_env.get_info()[0]['rgb']
+        return None
